+++ heat/cluster/tests/test_kmedoids.py
@@ -74,10 +74,10 @@ class TestKMeans(TestCase):
         split = 0
         # get some test data
         iris = ht.load("heat/datasets/iris.csv", sep=";", split=split)
-
+        ht.random.seed(1)
         # fit the clusters
         k = 3
-        kmedoid = ht.cluster.KMedoids(n_clusters=k)
+        kmedoid = ht.cluster.KMedoids(n_clusters=k, random_state=1)
         kmedoid.fit(iris)
 
         # check whether the results are correct
@@ -123,6 +123,10 @@ class TestKMeans(TestCase):
         kmedoid.fit(data)
         self.assertIsInstance(kmedoid.cluster_centers_, ht.DNDarray)
         self.assertEqual(kmedoid.cluster_centers_.shape, (4, 3))
+        for i in range(kmedoid.cluster_centers_.shape[0]):
+            self.assertTrue(
+                ht.any(ht.sum(ht.abs(kmedoid.cluster_centers_[i, :] - data), axis=1) == 0)
+            )
 
         # More Samples
         n = 100 * ht.MPI_WORLD.size
@@ -150,7 +154,10 @@ class TestKMeans(TestCase):
         self.assertEqual(kmedoid.cluster_centers_.shape, (4, 3))
         for i in range(kmedoid.cluster_centers_.shape[0]):
             self.assertTrue(
-                ht.any(ht.sum(ht.abs(kmedoid.cluster_centers_[i, :] - data), axis=1) == 0)
+                ht.any(
+                    ht.sum(ht.abs(kmedoid.cluster_centers_[i, :] - data.astype(ht.float32)), axis=1)
+                    == 0
+                )
             )
 
         # on Ints (different radius, offset and datatype