+++ heat/nn/data_parallel.py
@@ -1,51 +1,37 @@
-import bisect
+"""
+This file is for the general data parallel neural network classes.
+"""
 import warnings
-
 import torch
+import torch.distributed
 import torch.nn as tnn
 
 from collections import OrderedDict
-from typing import Callable, List, Union, Tuple
-from ..core.communication import MPICommunication
-from ..core.communication import MPI
-from .. import optim
-from ..core.devices import get_device
+from typing import Any, Callable, Dict, List, Union, Tuple
 
-__all__ = ["DataParallel"]
+from .. import optim
+from ..core.communication import MPI
+from ..core.communication import MPI_WORLD
+from ..core.communication import MPICommunication
+
+
+__all__ = ["DataParallel", "DataParallelMultiGPU"]
 
 
 class DataParallel(tnn.Module):
     """
     Implements data parallelism across multiple processes. This means that the same model will be run locally
-    on each process. Creation of the model parallels to PyTorch, the only changes are using HeAT layers (ht.nn.layer)
-    in the initialization of the network. I.E. (example code contains). If there is not a HeAT layer,
-    it will fall back to the PyTorch layer of the same name.
-    .. code-block:: python
+    on each process. Creation of the model is similar to PyTorch, the only changes are using HeAT layers (ht.nn.layer)
+    in the initialization of the network/optimizer. If there is not a HeAT layer, it will fall back to the PyTorch layer
+    of the same name. The same is true for the optimizer. It's possible to use more than one optimizer, but
+    communication during parameter updates is limited to blocking. The same limitation takes effect when passing an
+    optimizer that does not deal exactly with the set of model's parameters. For the given model both the
+    ``__init__()`` and ``forward()`` functions must be defined in the class defining the network.
 
-        class TestingModel(torch.nn.Module):
-            def __init__(self):
-                super(TestingModel, self).__init__()
-                self.net1 = ht.nn.Linear(10, 10)
-                self.relu = ht.nn.ReLU()
-                self.net2 = ht.nn.Linear(10, 5)
+    An example of this is shown in `examples/mnist.py <https://github.com/helmholtz-analytics/heat/blob/504-docstring-formatting/examples/nn/mnist.py>`_.
 
-            def forward(self, x):
-                return self.net2(self.relu(self.net1(x)))
-
-        t_model = TestingModel()
-        t_optimizer = torch.optim.SGD(t_model.parameters(), lr=0.01)
-        ht_optimizer = ht.optim.DataParallelOptimizer(t_optimizer)
-        ht_model = ht.nn.DataParallel(t_model, comm, ht_optimizer)
-
-    and a requirement of giving a HeAT communicator (``comm``, :class:`..core.communication.MPICommunication`)
-    and at least one DataParallelOptimizer (``dp_optimizers``, :class:`..optim.dp_optimizer.DataParallelOptimizer`).
-    It's possible to pass more than one optimizer, but communication during parameter updates is limited to blocking
-    then. The same limitation takes effect when passing an optimizer that does not deal exactly with the set of model's
-    parameters. For the given model both the ``__init__()`` and ``forward()`` functions must be defined in the class
-    defining the network.
-
-    It is highly recommended that a HeAT DataLoader is used, see :func:`..utils.data.datatools.DataLoader`. The
-    default communications scheme for this is blocking. The blocking scheme will average the model parameters during
+    It is highly recommended that a HeAT DataLoader is used, see :func:`ht.utils.data.DataLoader <heat.utils.data.datatools.DataLoader>`.
+    The default communications scheme for this is blocking. The blocking scheme will average the model parameters during
     the backwards step, synchronizing them before the next model iteration.
 
     Usage of more than one optimizer forces MPI communication to be parameter updates to use blocking communications.
@@ -69,7 +55,11 @@ class DataParallel(tnn.Module):
         comm: MPICommunication,
         optimizer: Union[optim.DataParallelOptimizer, List, Tuple],
         blocking_parameter_updates: bool = False,
-    ):
+    ):  # noqa: D107
+        if isinstance(optimizer, optim.DASO):
+            raise TypeError(
+                "For use with DASO please use DataParallelMultiGPU instead of DataParallel"
+            )
         super(DataParallel, self).__init__()
         self.module = module
         self.comm = comm
@@ -138,8 +128,11 @@ class DataParallel(tnn.Module):
                 param.register_hook(self._nonblocking_hook(layer_name, name))
         self._param_slices[layer_name_prev] = slice(start_idx, len(self._param_indices))
 
-    def __setattr__(self, name, value):
-        # auto-detect end of epoch's training phase and finalize wait handles (only relevant for non-blocking)
+    def __setattr__(self, name: str, value: Union[torch.nn.Module, torch.Tensor, Any]) -> None:
+        """
+        Overwrite the current torch.nn.Module.__setattr__ so that it auto-detects the end of epoch's
+        training phase and finalize wait handles (only relevant for non-blocking)
+        """
         if name == "training" and not value and not self.blocking_parameter_updates:
             self._iparam_update()
         super(DataParallel, self).__setattr__(name, value)
@@ -180,16 +173,16 @@ class DataParallel(tnn.Module):
         return ret
 
     def _iparam_update(self, param_slice: slice = None, layer_names: List[str] = None) -> None:
-        """
+        r"""
         Update parameters asynchronously via wait handles.
 
         Parameters
         ----------
         param_slice : slice, optional
-            Slice object for creating a view onto optimizer's params list.
+            Slice object for creating a view onto optimizer's params list.\n
             By default, the whole params list is used, (``None``)
         layer_names : list(str), optional
-            List of layer names which parameters will be updated, must match param_slice.
+            List of layer names which parameters will be updated, must match param_slice.\n
             By default, all layers are updated (``None``)
         """
         # for non-blocking, only one dp optimizer is allowed
@@ -209,16 +202,18 @@ class DataParallel(tnn.Module):
             if layer_name not in self._active_layers:
                 return
             # iterate over layer's parameters/associated wait handles
-            for (_, param_name, wait_handle) in self._layer_wait_handles[layer_name]:
+            for param_name, wait_handle, dtp, tens in self._layer_wait_handles[layer_name]:
                 # get internal index of selected parameter
                 param_idx = self._param_indices[param_name]
                 # synchronize, get parameter's global gradient
                 wait_handle.wait()
                 # check if shapes are matching
-                if dp_optimizer.params_ref[param_idx].grad.data.shape != wait_handle.tensor.shape:
+                if (
+                    dp_optimizer.params_ref[param_idx].grad.data.shape != tens.shape
+                ):  # wait_handle.tensor.shape:
                     raise ValueError("Shapes must be equal.")
                 # accumulate parameter's global gradient
-                dp_optimizer.params_ref[param_idx].grad.data += wait_handle.tensor
+                dp_optimizer.params_ref[param_idx].grad.data += tens.to(dtp)  # wait_handle.tensor
                 # remove layer from set of active layers, if present
                 self._active_layers.discard(layer_name)
         # if desired, perform actual parameter update
@@ -238,11 +233,12 @@ class DataParallel(tnn.Module):
         ----------
         [1] (cf. https://pytorch.org/docs/stable/tensors.html#torch.Tensor.register_hook).
         """
+        grad_loc_bf = grad_loc.to(torch.float)  # bfloat16)
         # average local gradients
-        grad_loc *= 1 / float(self.comm.size)
+        grad_loc_bf *= 1 / float(self.comm.size)
         # perform MPI Allreduce to compute global gradient
-        self.comm.Allreduce(MPI.IN_PLACE, grad_loc, MPI.SUM)
-        return grad_loc
+        self.comm.Allreduce(MPI.IN_PLACE, grad_loc_bf, MPI.SUM)  # mpi_sum_bf16)
+        return grad_loc_bf.to(grad_loc.dtype)
 
     def _nonblocking_hook(self, layer_name: str, param_name: str) -> Callable:
         """
@@ -257,23 +253,27 @@ class DataParallel(tnn.Module):
         """
         # hook function for blocking gradient data exchange
         def _hook(grad_loc: torch.Tensor) -> torch.Tensor:
+            with torch.no_grad():
+                wrk = grad_loc.to(torch.float)  # bfloat16)
             # counterbalance local gradient averaging
-            grad_loc *= 1 / float(self.comm.size)
+            wrk *= 1 / float(self.comm.size)
             # perform MPI IAllreduce to compute global gradient, returns wait handle
-            wait_handle = self.comm.Iallreduce(MPI.IN_PLACE, grad_loc, MPI.SUM)
+            wait_handle = self.comm.Iallreduce(MPI.IN_PLACE, wrk, MPI.SUM)  # mpi_sum_bf16)
             # if layer wait handle dict does not contain the layer, add it -> automatically tracks reversed layer order
             if layer_name not in self._layer_wait_handles:
                 self._layer_wait_handles[layer_name] = list()
             # add layer to set of active layers
             self._active_layers.add(layer_name)
             # assign wait handle to its layer, layer-internal sorting by size
-            bisect.insort(
-                self._layer_wait_handles[layer_name], (grad_loc.numel(), param_name, wait_handle)
-            )
+            # bisect.insort(
+            #     self._layer_wait_handles[layer_name], (wrk.numel(), param_name, wait_handle)
+            # )
             # TODO: is sorting faster? or is there any difference?
-            # self._layer_wait_handles[layer_name].append((grad_loc.numel(), param_name, wait_handle))
+            self._layer_wait_handles[layer_name].append(
+                (param_name, wait_handle, grad_loc.dtype, wrk)
+            )
             # don't return grad_loc, otherwise gradient is doubled
-            return torch.zeros(*grad_loc.size(), device=get_device().torch_device)
+            return torch.zeros(*wrk.size(), device=grad_loc.device)
 
         return _hook
 
@@ -295,6 +295,71 @@ class DataParallel(tnn.Module):
             return input_
 
         return _hook
+
+    @staticmethod
+    def _reset_parameters(module: tnn.Module) -> None:
+        """
+        Reset parameters of given torch submodule. Only works for basic module types containing ``reset_parameters``
+        function.
+
+        Parameters
+        ----------
+        module: torch.nn.Module
+            Submodule whose parameters are to be reset
+        """
+        if callable(getattr(module, "reset_parameters", None)):
+            module.reset_parameters()
+
+
+class DataParallelMultiGPU(tnn.Module):
+    """
+    This creates data parallel networks local to each node using PyTorch's distributed class. This does NOT
+    do any global synchronizations. To make optimal use of this structure, use :func:`ht.optim.DASO <heat.optim.dp_optimizer.DASO>`.
+
+    Notes
+    -----
+    The PyTorch distributed process group must already exist before this class is initialized.
+
+    Parameters
+    ----------
+    module: torch.nn.Module
+        an implemented PyTorch model
+    optimizer: optim.DASO
+        A DASO optimizer. Other optimizers are not yet implemented. The DASO optimizer should be
+        defined prior to calling this class.
+    comm: MPICommunication, optional
+        A global communicator.
+        Default: :func:`MPICommunication <heat.core.comm.MPICommunication>`
+    """
+
+    def __init__(
+        self, module: torch.nn.Module, optimizer: optim.DASO, comm: MPICommunication = MPI_WORLD
+    ):  # noqa: D107
+        super(DataParallelMultiGPU, self).__init__()
+        rank = comm.rank
+        if torch.cuda.device_count() > 1:
+            self.loc_gpus = torch.cuda.device_count()
+            local_rank = rank % self.loc_gpus
+            device = "cuda:" + str(local_rank)
+            torch.cuda.set_device(device=device)
+            module = tnn.parallel.DistributedDataParallel(module, device_ids=[local_rank])
+        else:
+            warnings.warn(
+                "DataParallelMultiGPU should be used with multiple GPUs per node", UserWarning
+            )
+        self.module = module
+        self.comm = comm
+
+        # unify parameters across nodes by unifying the random seed and resetting parameters
+        self.module.apply(self._reset_parameters)
+
+        optimizer.set_model(self.module)
+
+    def forward(self, *inputs: Tuple, **kwargs: Dict) -> torch.Tensor:
+        """
+        Calls the forward method for the torch model
+        """
+        return self.module(*inputs, **kwargs)
 
     @staticmethod
     def _reset_parameters(module: tnn.Module) -> None: